server/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
sqlalchemy==2.0.38
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
langchain==0.3.20
langchain-community==0.3.19
langchain-text-splitters==0.3.6
langchain-chroma==0.2.2
chromadb==0.4.22
sentence-transformers==2.5.1
pypdf==4.0.2
docx2txt==0.8
llama-cpp-python==0.2.56
pydantic-settings==2.8.1
openai==1.65.4
